requirements.txt
transformers>=4.35.0
torch>=2.0.0
langdetect>=1.0.9
spacy>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
indic-nlp-library>=0.91
huggingface-hub>=0.19.0
sentence-transformers>=2.2.0
flair>=0.12.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
